backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "aiomysql>=0.3.2",
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pre-commit>=4.6.0",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "ruff>=0.15.11",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.45.0",
]
